feat: move '/' to it's own folder called 'home'
