+++ src/app/globals.css
@@ -1,6 +1,5 @@
 body{
   margin: 0;
-  /* background: linear-gradient(171deg, rgb(90, 150, 218) 0%, rgba(204,204,204,1) 100%);  */
   height: 100vh;
   text-align: center;
 }
