+++ src/app/home/page.module.css
@@ -1,0 +1,57 @@
+.textarea {
+  margin-top: 20px;
+  margin-bottom: 5px;
+}
+
+.button {
+  margin-bottom: 20px;
+  height: 30px;
+  width: 90px;
+}
+
+.post{
+  background-color: white;
+  border-radius: 25px;
+  display: grid;
+  grid-template-columns: 30% 70%;
+  gap: 10px;
+  margin-bottom: 15px;
+  font-family: 'Trebuchet MS', 'Lucida Sans Unicode', 'Lucida Grande', 'Lucida Sans', Arial, sans-serif;
+  text-align: left;
+  padding-left: 5px;
+}
+
+.pic {
+  grid-column: 1/2;
+  grid-row: 1/3;
+  border-radius: 25px;
+  border: 3px solid yellow;
+  padding-top: 20px;
+  margin-top: 10px;
+}
+
+.image {
+  width: 75px;
+  height: 75px;
+  border: 3 solid green;
+}
+
+.time {
+ grid-column: 2/2;
+ grid-row: 2/2;
+ text-align: left;
+ padding-bottom: 10px;
+ color: gray;
+}
+
+.user {
+ padding-top: 10px;
+ font-weight: bold;
+ text-align: left;
+}
+
+.message {
+grid-row: 3/3;
+grid-column: 1/3;
+padding-bottom: 10px;
+}
