+++ src/app/landing/landing.module.css
@@ -38,8 +38,6 @@
   font-size: 50px
 }
 
-
-
 .area {
   background: #767676;
   background: -webkit-linear-gradient(to left, #8f94fb, #4e54c8);
